Move particle scripting class into separate file
